+++ include/visionaray/detail/cpu_buffer_rt.inl
@@ -115,7 +115,7 @@ void cpu_buffer_rt<ColorFormat, DepthFormat>::end_frame()
 }
 
 template <pixel_format ColorFormat, pixel_format DepthFormat>
-void cpu_buffer_rt<ColorFormat, DepthFormat>::resize(size_t w, size_t h)
+void cpu_buffer_rt<ColorFormat, DepthFormat>::resize(int w, int h)
 {
     render_target::resize(w, h);
