+++ include/visionaray/detail/gpu_buffer_rt.inl
@@ -81,7 +81,7 @@ void gpu_buffer_rt<ColorFormat, DepthFormat>::end_frame()
 }
 
 template <pixel_format ColorFormat, pixel_format DepthFormat>
-void gpu_buffer_rt<ColorFormat, DepthFormat>::resize(size_t w, size_t h)
+void gpu_buffer_rt<ColorFormat, DepthFormat>::resize(int w, int h)
 {
     render_target::resize(w, h);
